+++ core/src/main/scala/com/github/j5ik2o/reactive/redis/feature/StringsFeature.scala
@@ -14,21 +14,21 @@ import scala.concurrent.duration.FiniteDuration
 trait StringsFeature {
   this: RedisClient =>
 
-  def append(key: String, value: String): ReaderTTaskRedisConnection[Result[Int]] =
+  def append(key: String, value: String): ReaderTTaskRedisConnection[Result[Long]] =
     send(AppendRequest(UUID.randomUUID(), key, value)).flatMap {
       case AppendSuspended(_, _)         => ReaderTTask.pure(Suspended)
       case AppendSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
       case AppendFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
     }
 
-  def bitCount(key: String, startAndEnd: Option[StartAndEnd] = None): ReaderTTaskRedisConnection[Result[Int]] =
+  def bitCount(key: String, startAndEnd: Option[StartAndEnd] = None): ReaderTTaskRedisConnection[Result[Long]] =
     send(BitCountRequest(UUID.randomUUID(), key, startAndEnd)).flatMap {
       case BitCountSuspended(_, _)         => ReaderTTask.pure(Suspended)
       case BitCountSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
       case BitCountFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
     }
 
-  def bitField(key: String, options: BitFieldRequest.SubOption*): ReaderTTaskRedisConnection[Result[Seq[Int]]] =
+  def bitField(key: String, options: BitFieldRequest.SubOption*): ReaderTTaskRedisConnection[Result[Seq[Long]]] =
     send(BitFieldRequest(UUID.randomUUID(), key, options: _*)).flatMap {
       case BitFieldSuspended(_, _)          => ReaderTTask.pure(Suspended)
       case BitFieldSucceeded(_, _, results) => ReaderTTask.pure(Provided(results))
@@ -38,7 +38,7 @@ trait StringsFeature {
   def bitOp(operand: BitOpRequest.Operand,
             outputKey: String,
             inputKey1: String,
-            inputKey2: String): ReaderTTaskRedisConnection[Result[Int]] =
+            inputKey2: String): ReaderTTaskRedisConnection[Result[Long]] =
     send(BitOpRequest(UUID.randomUUID(), operand, outputKey, inputKey1, inputKey2)).flatMap {
       case BitOpSuspended(_, _)         => ReaderTTask.pure(Suspended)
       case BitOpSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
@@ -47,20 +47,20 @@ trait StringsFeature {
 
   def bitPos(key: String,
              bit: Int,
-             startAndEnd: Option[BitPosRequest.StartAndEnd] = None): ReaderTTaskRedisConnection[Result[Int]] =
+             startAndEnd: Option[BitPosRequest.StartAndEnd] = None): ReaderTTaskRedisConnection[Result[Long]] =
     send(BitPosRequest(UUID.randomUUID(), key, bit, startAndEnd)).flatMap {
       case BitPosSuspended(_, _)         => ReaderTTask.pure(Suspended)
       case BitPosSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
       case BitPosFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
     }
 
-  def decr(key: String): ReaderTTaskRedisConnection[Result[Int]] = send(DecrRequest(UUID.randomUUID(), key)).flatMap {
+  def decr(key: String): ReaderTTaskRedisConnection[Result[Long]] = send(DecrRequest(UUID.randomUUID(), key)).flatMap {
     case DecrSuspended(_, _)         => ReaderTTask.pure(Suspended)
     case DecrSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
     case DecrFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
   }
 
-  def decrBy(key: String, value: Int): ReaderTTaskRedisConnection[Result[Int]] =
+  def decrBy(key: String, value: Int): ReaderTTaskRedisConnection[Result[Long]] =
     send(DecrByRequest(UUID.randomUUID(), key, value)).flatMap {
       case DecrBySuspended(_, _)         => ReaderTTask.pure(Suspended)
       case DecrBySucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
@@ -74,7 +74,7 @@ trait StringsFeature {
       case GetFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
     }
 
-  def getBit(key: String, offset: Int): ReaderTTaskRedisConnection[Result[Int]] =
+  def getBit(key: String, offset: Int): ReaderTTaskRedisConnection[Result[Long]] =
     send(GetBitRequest(UUID.randomUUID(), key, offset)).flatMap {
       case GetBitSuspended(_, _)        => ReaderTTask.pure(Suspended)
       case GetBitSucceeded(_, _, value) => ReaderTTask.pure(Provided(value))
@@ -95,13 +95,13 @@ trait StringsFeature {
       case GetSetFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
     }
 
-  def incr(key: String): ReaderTTaskRedisConnection[Result[Int]] = send(IncrRequest(UUID.randomUUID(), key)).flatMap {
+  def incr(key: String): ReaderTTaskRedisConnection[Result[Long]] = send(IncrRequest(UUID.randomUUID(), key)).flatMap {
     case IncrSuspended(_, _)         => ReaderTTask.pure(Suspended)
     case IncrSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
     case IncrFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
   }
 
-  def incrBy(key: String, value: Int): ReaderTTaskRedisConnection[Result[Int]] =
+  def incrBy(key: String, value: Int): ReaderTTaskRedisConnection[Result[Long]] =
     send(IncrByRequest(UUID.randomUUID(), key, value)).flatMap {
       case IncrBySuspended(_, _)         => ReaderTTask.pure(Suspended)
       case IncrBySucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
@@ -150,7 +150,7 @@ trait StringsFeature {
       case SetFailed(_, _, ex) => ReaderTTask.raiseError(ex)
     }
 
-  def setBit(key: String, offset: Int, value: Int): ReaderTTaskRedisConnection[Result[Int]] =
+  def setBit(key: String, offset: Int, value: Int): ReaderTTaskRedisConnection[Result[Long]] =
     send(SetBitRequest(UUID.randomUUID(), key, offset, value)).flatMap {
       case SetBitSuspended(_, _)         => ReaderTTask.pure(Suspended)
       case SetBitSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
@@ -164,14 +164,14 @@ trait StringsFeature {
       case SetExFailed(_, _, ex) => ReaderTTask.raiseError(ex)
     }
 
-  def setRange[A: Show](key: String, range: Int, value: A): ReaderTTaskRedisConnection[Result[Int]] =
+  def setRange[A: Show](key: String, range: Int, value: A): ReaderTTaskRedisConnection[Result[Long]] =
     send(SetRangeRequest(UUID.randomUUID(), key, range, value)).flatMap {
       case SetRangeSuspended(_, _)         => ReaderTTask.pure(Suspended)
       case SetRangeSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
       case SetRangeFailed(_, _, ex)        => ReaderTTask.raiseError(ex)
     }
 
-  def strLen(key: String): ReaderTTaskRedisConnection[Result[Int]] =
+  def strLen(key: String): ReaderTTaskRedisConnection[Result[Long]] =
     send(StrLenRequest(UUID.randomUUID(), key)).flatMap {
       case StrLenSuspended(_, _)         => ReaderTTask.pure(Suspended)
       case StrLenSucceeded(_, _, result) => ReaderTTask.pure(Provided(result))
