+++ core/src/main/scala/com/github/j5ik2o/reactive/redis/parser/model/NumberExpr.scala
@@ -1,3 +1,3 @@
 package com.github.j5ik2o.reactive.redis.parser.model
 
-final case class NumberExpr(value: Int) extends Expr
+final case class NumberExpr(value: Long) extends Expr
